% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_risk.R
\name{temp_risk}
\alias{temp_risk}
\title{temp_risk}
\usage{
temp_risk(
  age,
  temperature,
  age_threshold_med = 50,
  age_threshold_high = 80,
  temperature_threshold = 40
)
}
\arguments{
\item{temperature}{(C)}

\item{age_threshold_med}{(years) (default 5)}

\item{age_threshold_high}{(years) (default 8)}

\item{temperature_threshold}{(C) (default 10)}

\item{threshold}{temp threshold (C) (or extreme temp)}
}
\value{
risk (high,med, low), mean nutrient concentration (mg/L)
}
\description{
compute the risk associated with high temperatures vs age
}
